requirements.txt
beautifulsoup4==4.4.1
click==6.6
colorama==0.3.7
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
et-xmlfile==1.0.1
gunicorn==19.6.0
httpie==0.9.3
jdcal==1.2
Markdown==2.6.6
mysqlclient==1.3.7
openpyxl==2.3.5
plotly==1.11.0
psycopg2==2.6.2
py==1.4.31
pygal==2.2.3
Pygments==2.1.3
pytest==2.9.1
pytz==2016.4
requests==2.10.0
script==0.1
six==1.10.0
static3==0.7.0
virtualenv==15.0.2
whitenoise==3.2
youtube-dl==2016.6.14
